+++ frontend/src/i18n/locales/source/ptBR.ts
@@ -115,7 +115,7 @@ export const ptBRMessages = {
   "views.admin.Account.clearInbox": "Limpar caixa de entrada",
   "views.index.AccountSettings.clearSentItems": "Limpar itens enviados",
   "views.admin.Account.clearSentItems": "Limpar itens enviados",
-  "views.Header.title": "Cloudflare E-mail Temporário",
+  "views.Header.title": "MiaoMail",
   "views.admin.DatabaseManager.code_db_version": "Versão do banco exigida pelo código",
   "views.user.UserOauth2Callback.codeNotFound": "código não encontrado",
   "views.admin.AccountSettings.config": "Configuração",